[CHORE] swipe gesture listener class 생성 #60
